+++ thoughtbubble-api/config/enviroment.ts
@@ -2,7 +2,6 @@ const dotenv = require('dotenv');
 dotenv.config();
 
 // purpose is to have intellisense for my env vars
-
 export const config = {
   port: process.env.PORT,
   node_env: process.env.NODE_ENV,
@@ -18,6 +17,9 @@ export const config = {
     github_client_id: process.env.GITHUB_CLIENT_ID,
     github_client_secret: process.env.GITHUB_CLIENT_SECRET,
     github_callback_url: process.env.GITHUB_CALLBACK_URL,
+    github_client_id_vscode: process.env.GITHUB_CLIENT_ID_VSCODE,
+    github_client_secret_vscode: process.env.GITHUB_CLIENT_SECRET_VSCODE,
+    github_callback_url_vscode: process.env.GITHUB_CALLBACK_URL_VSCODE,
   },
   vscode_dev: {
     vscode_dev_token: process.env.VSCODE_DEV_TOKEN,
